Move conversion hosts request to general step, use selector
