+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardClustersStep/MappingWizardClustersStep.js
@@ -6,31 +6,20 @@ import { length } from 'redux-form-validators';
 
 import ClustersStepForm from './components/ClustersStepForm/ClustersStepForm';
 
-import { OPENSTACK_CONVERSION_HOST_TYPE } from './MappingWizardClustersStepConstants';
-
 class MappingWizardClustersStep extends React.Component {
   componentDidMount() {
+    const { targetProvider, ospConversionHosts, showAlertAction } = this.props;
     this.fetchClusters();
-  }
-
-  componentDidUpdate(prevProps) {
-    const { targetProvider, conversionHosts, showAlertAction } = this.props;
-    if (targetProvider === 'openstack' && !prevProps.conversionHosts && conversionHosts) {
-      const openstackConversionHosts = conversionHosts.filter(
-        host => host.resource.type === OPENSTACK_CONVERSION_HOST_TYPE
+    if (targetProvider === 'openstack' && ospConversionHosts.length === 0) {
+      showAlertAction(
+        __('At least one host must be enabled as a conversion host. You can continue to create an infrastructure mapping, but you must configure a conversion host before migration execution.'), // prettier-ignore
+        'warning'
       );
-      if (openstackConversionHosts.length === 0) {
-        showAlertAction(
-          __('At least one host must be enabled as a conversion host. You can continue to create an infrastructure mapping, but you must configure a conversion host before migration execution.'), // prettier-ignore
-          'warning'
-        );
-      }
     }
   }
 
   componentWillUnmount() {
     this.props.hideAlertAction();
-    this.props.clearConversionHostsAction();
   }
 
   fetchClusters = () => {
@@ -39,8 +28,6 @@ class MappingWizardClustersStep extends React.Component {
       fetchSourceClustersAction,
       fetchTargetComputeUrls,
       fetchTargetClustersAction,
-      fetchConversionHostsUrl,
-      fetchConversionHostsAction,
       targetProvider,
       queryHostsUrl,
       queryHostsAction
@@ -64,10 +51,6 @@ class MappingWizardClustersStep extends React.Component {
         queryHostsAction(queryHostsUrl, hostIDsByClusterID);
       }
     });
-
-    if (targetProvider === 'openstack') {
-      fetchConversionHostsAction(fetchConversionHostsUrl);
-    }
   };
 
   render() {
@@ -123,9 +106,6 @@ MappingWizardClustersStep.propTypes = {
   fetchSourceClustersAction: PropTypes.func,
   fetchTargetComputeUrls: PropTypes.object,
   fetchTargetClustersAction: PropTypes.func,
-  fetchConversionHostsAction: PropTypes.func,
-  fetchConversionHostsUrl: PropTypes.string,
-  clearConversionHostsAction: PropTypes.func,
   queryHostsUrl: PropTypes.string,
   queryHostsAction: PropTypes.func,
   showAlertAction: PropTypes.func,
@@ -140,7 +120,7 @@ MappingWizardClustersStep.propTypes = {
   isFetchingHostsQuery: PropTypes.bool,
   isRejectedHostsQuery: PropTypes.bool,
   hostsByClusterID: PropTypes.object,
-  conversionHosts: PropTypes.array
+  ospConversionHosts: PropTypes.array
 };
 MappingWizardClustersStep.defaultProps = {
   fetchSourceClustersAction: noop,
@@ -155,7 +135,7 @@ MappingWizardClustersStep.defaultProps = {
   isFetchingHostsQuery: false,
   isRejectedHostsQuery: false,
   hostsByClusterID: {},
-  conversionHosts: null,
+  ospConversionHosts: [],
   fetchSourceClustersUrl:
     '/api/clusters?expand=resources' +
     '&attributes=ext_management_system.emstype,v_parent_datacenter,ext_management_system.name' +
@@ -166,8 +146,7 @@ MappingWizardClustersStep.defaultProps = {
       '&attributes=ext_management_system.emstype,v_parent_datacenter,ext_management_system.name,hosts' +
       '&filter[]=ext_management_system.emstype=rhevm',
     openstack: '/api/cloud_tenants?expand=resources&attributes=ext_management_system.name,ext_management_system.id'
-  },
-  fetchConversionHostsUrl: '/api/conversion_hosts?attributes=resource&expand=resources'
+  }
 };
 
 export default reduxForm({
