+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardClustersStep/MappingWizardClustersStepConstants.js
@@ -1,7 +1,5 @@
 export const FETCH_V2V_SOURCE_CLUSTERS = 'FETCH_V2V_SOURCE_CLUSTERS';
 export const FETCH_V2V_TARGET_CLUSTERS = 'FETCH_V2V_TARGET_CLUSTERS';
 export const QUERY_V2V_HOSTS = 'QUERY_V2V_HOSTS';
-export const FETCH_V2V_CONVERSION_HOSTS = 'FETCH_V2V_CONVERSION_HOSTS';
-export const CLEAR_V2V_CONVERSION_HOSTS = 'CLEAR_V2V_CONVERSION_HOSTS';
 
 export const OPENSTACK_CONVERSION_HOST_TYPE = 'ManageIQ::Providers::Openstack::CloudManager::Vm';
