+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardClustersStep/MappingWizardClustersStepSelectors.js
@@ -1,3 +1,5 @@
+import { OPENSTACK_CONVERSION_HOST_TYPE } from './MappingWizardClustersStepConstants';
+
 export const sourceClustersFilter = (sourceClustersToFilter, clustersStepMappings) => {
   const mappedSourceClusters = clustersStepMappings.reduce(
     (sourceClusters, targetClusterWithSourceClusters) => sourceClusters.concat(targetClusterWithSourceClusters.nodes),
@@ -8,3 +10,6 @@ export const sourceClustersFilter = (sourceClustersToFilter, clustersStepMapping
     sourceCluster => !mappedSourceClusters.some(mappedSourceCluster => mappedSourceCluster.id === sourceCluster.id)
   );
 };
+
+export const ospConversionHostsFilter = conversionHosts =>
+  conversionHosts.filter(host => host.resource.type === OPENSTACK_CONVERSION_HOST_TYPE);
