+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardGeneralStep/MappingWizardGeneralStepConstants.js
@@ -1,1 +1,2 @@
 export const SET_V2V_EDITING_MAPPING = 'SET_V2V_EDITING_MAPPING';
+export const FETCH_V2V_CONVERSION_HOSTS = 'FETCH_V2V_CONVERSION_HOSTS';
